requirements.txt
Flask==3.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
SQLAlchemy==2.0.36
Werkzeug==3.1.3
psycopg2==2.9.10
pandas==2.2.3
PyMuPDF==1.24.14
google-api-python-client==2.151.0
google-auth==2.36.0
Pillow==11.0.0